add anonim name columns

+++ app/Http/Livewire/Guest/Sambat/ModalDelete.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire\Guest\Sambat;
 
+use App\Constants\AppPermissions;
 use LivewireUI\Modal\ModalComponent;
 use App\Models\Sambat;
 use Illuminate\Support\Facades\Storage;
@@ -12,13 +13,16 @@ class ModalDelete extends ModalComponent
 
     public function handleForm()
     {
+        $sambat = Sambat::find($this->sambat_id);
+        if (auth()->id() != $sambat->user_id and !auth()->user()->can(AppPermissions::DELETE_SAMBAT))
+            return $this->emit('error', "Gagal menghapus sambat");
+
         try {
-            $sambat = Sambat::find($this->sambat_id);
             $sambat->tags()->detach();
             $sambat->comments()->delete();
             $sambat->votes()->delete();
-
             foreach ($sambat->images as $image) Storage::disk('public')->delete($image->url);
+            $sambat->images()->delete();
             $sambat->delete();
 
             $this->emit('success', "Sukses menghapus sambat");
